refactor: move sorting logic to a dedicated sortCollectionItems function and add comprehensive unit tests
